pyproject.toml
[project]
name = "web-builder-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.39.2",
    "progress>=1.6.1",
    "python-dotenv>=1.1.1",
    "randomname>=0.2.1",
    "strands-agents>=0.2.0",
    "strands-agents-tools>=0.1.8",
]
